requirements.txt
discord.py
json
nltk
numpy
pickle
requests
requests-cache
tensorflow
tflearn
